Added sync service project

+++ Cadenza/Cadenza.Local/Services/FileAccess.cs
@@ -11,28 +11,40 @@ public class FileAccess : IFileAccess
 
     public void SaveText(string path, string text)
     {
+        var directory = Path.GetDirectoryName(path);
+        Directory.CreateDirectory(directory);
         File.WriteAllText(path, text);
     }
 
-    public IEnumerable<LocalFile> GetFiles(string directoryPath, List<string> extensions)
+    public List<LocalFile> GetFiles(string directoryPath, List<string> extensions)
     {
+        if (!Directory.Exists(directoryPath))
+            return new List<LocalFile>();
+
+        var files = new List<LocalFile>();
+
         foreach (var filepath in Directory.GetFiles(directoryPath, "*", SearchOption.AllDirectories))
         {
             var fileInfo = new FileInfo(filepath);
             if (extensions.Contains(fileInfo.Extension))
             {
-                yield return new LocalFile
+                files.Add(new LocalFile
                 {
                     Path = fileInfo.FullName,
                     DateCreated = fileInfo.CreationTime,
                     DateModified = fileInfo.LastWriteTime
-                };
+                });
             }
         }
+
+        return files;
     }
 
     public void DeleteFile(string path)
     {
+        if (!File.Exists(path))
+            return;
+
         File.Delete(path);
     }
 }
